# openapi.yaml
openapi: 3.0.3
info:
  title: Redstick Ventures Command Center API
  description: |
    Complete API for the Redstick Ventures Command Center platform.
    
    ## Authentication
    All API requests require a JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    - Authenticated: 1000 requests/minute
    - Unauthenticated: 20 requests/minute
    
    ## Base URLs
    - Production: https://api.redstick.vc
    - Staging: https://api-staging.redstick.vc
    - Development: http://localhost:3000/api
  version: 1.0.0
  contact:
    name: Redstick API Support
    email: api-support@redstick.vc
    url: https://docs.redstick.vc
  license:
    name: Proprietary
    url: https://redstick.vc/terms
servers:
  - url: https://api.redstick.vc/api/v1
    description: Production
  - url: https://api-staging.redstick.vc/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Development
security:
  - bearerAuth: []
paths:
...
  /companies/{id}:
    get:
      tags:
        - Companies
      summary: Get a specific company
      parameters:
        - $ref: '#/components/parameters/CompanyIdParam'
      responses:
        '200':
          description: Company details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Companies
      summary: Update a company
      parameters:
        - $ref: '#/components/parameters/CompanyIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyRequest'
      responses:
        '200':
          description: Company updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'

    delete:
      tags:
        - Companies
      summary: Delete a company
      parameters:
        - $ref: '#/components/parameters/CompanyIdParam'
      responses:
        '200':
          description: Company deleted

  # Agents
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login
  parameters:
...
    CompanyIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Company ID
...
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          type: object

    # Pagination
...
    Company:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        website:
          type: string
          format: uri
        industry:
          type: string
        subIndustry:
          type: string
        stage:
          type: string
        foundedYear:
          type: integer
        employees:
          type: integer
        location:
          type: object
          properties:
            city:
              type: string
            state:
              type: string
            country:
              type: string
        socialLinks:
          type: object
          properties:
            linkedin:
              type: string
            twitter:
              type: string
        founders:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              title:
                type: string
              linkedin:
                type: string
        fundingHistory:
          type: array
          items:
            type: object
        metrics:
          type: object
          properties:
            revenue:
              type: number
            growthRate:
              type: number
            customers:
              type: integer
        dealCount:
          type: integer
        activeDealId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CompanyResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Company'
...
    CompanyDetailResponse:
      allOf:
        - $ref: '#/components/schemas/CompanyResponse'
    CreateCompanyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        website:
          type: string
        industry:
          type: string
        subIndustry:
          type: string
        stage:
          type: string
        foundedYear:
          type: integer
        employees:
          type: integer
        location:
          type: object
          properties:
            city:
              type: string
            state:
              type: string
            country:
              type: string
        socialLinks:
          type: object
        founders:
          type: array
          items:
            type: object
    UpdateCompanyRequest:
      allOf:
        - $ref: '#/components/schemas/CreateCompanyRequest'
      type: object

    # Agent
...
    Error:
      type: object
      required:
        - success
        - error
        - code
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
        details:
          type: object
        requestId:
          type: string
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
            requestId: "req_ghi789"
...